Make local token provider from value configurable
